requirements.txt
pytest>=7.0
coverage>=7.0
flake8>=6.0
torch>=2.5.1
pyyaml>=6.0
numpy>=1.24
sbi>=0.21.0
pandas>=2.2.3
matplotlib>=3.10.1
seaborn>=0.13.2
scikit-learn~=1.6.1
hydra-core>=1.3
omegaconf>=2.0
